[lipstick] Make application blacklisting more dynamic. Contributes to JB#37469
Now apps can be made blacklisted during runtime. LauncherModel will
contain always also blacklisted apps but those are not populated
to the LauncherFolderModel. Blacklist applications change triggers
app blacklisting meaning that apps that matches to the change are
removed from the LauncherFolderModel. Similarly
when blacklisted apps changes and an app is no longer blacklisted,
the app is brought back to the LauncherFolderModel to the index where
it should be.

Saved menu xml contains always also blacklisted apps.