[notifications] Suppress display wakeup for usb cable disconnect. JB#37024
Powering up display when user detaches usb cable just to show a banner about
cable disconnect is a bit excessive.

Suppress display wakeups for device.removed notification category. This way
the banner will be show if cable is detached while display is already on,
but the display is not powered up just because of the disconnect.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>