[lipstick] Write screenshots to disk asynchronously. Contributes to JB#42583
It can take a significant amount of time to save the image on devices
with larger screens. Do the save in a thread and return an object
which can notify when it's done.  The HomeApplication API which returns
a boolean indicating success continues to block, wouldn't want a returned
object to be mistaken for success.