[lipstick] Add configuration options for the application .desktop directories. Contributes to JB#43964
The /desktop/sailfish/application_desktop_paths value provides the
default directories for application .desktop files. If empty the
list will default to the xdg directories. The
/desktop/sailfish/url_desktop_paths value provides additional
directories from which only .desktop files of the URL type are
read. If empty it will default to the xdg directories not in
application_desktop_paths.