Add lipstick-specific VPN Agent documentation
The main VPN Agent documentation can be found in
connman/doc/vpn-agent-api.txt, which details the API used by connman to
contact lipstick to request user credentials (triggering lipstick to
display a dialogue to be filled out by the user).

Lipstick provides some soft-extensions to the protocol, and this
documentation explains them.