[lipstick] Add option to keep VPN agent creds. Contributes to JB#47237
Implemented option "KeepCredentials" into VPN agent to make it
possible to prevent credential removal even though storage and retrieval
is forbidden. An example case of this is the OpenVPN Encrypted Private
Key password, which is requested in second prompt after the credentials.
The password is not to be stored by VPN agent but is kept in memory and
for this reason storage and retrieval are set to be false, but the
credentials should not be cleared when the Private Key password is
requested.

This change answers to that need. By default "KeepCredentials" is
set to be false and is not stored in settings. But when explictly set as
true, it will be recorded as "keepCredentials" and later used when
creating response to the request. In the VpnAgent:respond() this value
is utilized to not to clear the connection credentials if not explicitly
set, as it defaults to false.