[lipstick] Read sandboxing mode from sailjaild. JB#54498 OMP#JOLLA-178
Call sailjaild's GetAppInfo to get sandboxing mode for apps. This is
needed to make lipstick prompt for sailjail when a legacy app is
sandboxed. In that case the information can not be read from the desktop
file. Without this change prompting is left to sailjail command which
also checks whether the app can be launched. This should also simplify
cover handling by allowing lipstick to manage when to show app cover
after the prompt is approved.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>