[lipstick] Suppress charging warnings when charging is disabled. Fixes JB#58812
When charging has been disabled for the purpose of limiting charging
time and/or cycles, battery level is expected to drop while charger is
connected but "Not enough power to charge" warning notifications are
still triggered.

Take charging enabled/disabled state into account and suppress warnings
while charging is disabled.

QMceChargingState is available in libmce-qt >= 1.5.0

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>