[compositor] Enablers for compositor handover debugging
There is some delay between having graphics stack initialized to the
point that something can be drawn and having application startup
progress to the point that the actual UI can be shown.

This delay can be visualized by setting DEBUG_COMPOSITOR_HANDOVER=y
in environment, in which case lipstick shows full screen magenta
window as soon as possible.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>