[lipstick] Do not change signal handlers on shutdown. JB#60078
Getting a shutdown notification basically means that we are going to
get a TERM signal in a couple of seconds. Using that as a cue for
substituting TERM handler that triggers lipstick to make a clean
exit with something else makes little sense. Especially as the
candidates are async-signal unsafe code from hwcomposer-plugin (that
can hang and require KILL) or default handler that makes quick and
dirty exit (potentially leaving behind e.g. lock files).

Keep using the already installed lipstick signal handlers.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>